+++ src/include/critter/Critter.h
@@ -14,7 +14,7 @@
 
 class Critter : public Observable, public FlowFieldObserver {
 public:
-	Critter(int level, float speed, int hitPoints, int strength, int reward, SDL_FRect start, Map* map);
+	Critter(int level, float speed, float hitPoints, int strength, int reward, SDL_FRect start, Map* map);
 	~Critter();
 
 	static constexpr float CRITTER_WIDTH_SCALE = 1.00f;
@@ -39,8 +39,8 @@ public:
 	int getReward() const { return reward; }
 	int getStrength() const { return strength; }
 
-	int getHitPoints() const { return hitPoints; }
-	void setHitPoints(int hitPoints);
+	float getHitPoints() const { return hitPoints; }
+	void setHitPoints(float hitPoints);
 
 	bool isClicked() const;
 	bool isDamaged() const { return isHurt; };
